requirements.txt
Flask
flask-cors
openai==0.28.1
pydub
requests
python-dotenv
elevenlabs==0.2.27
google-cloud-texttospeech
google-cloud-speech
google-cloud-storage
python-docx
